Initial commit
VS uses UTF-16 for its text buffers and Nvim normally uses UTF-8,
however, to reduce memory usage and to avoid the need for
synchronization Nvim has been patched to use VS buffers. Each time Nvim
reads or modifies text in the buffer, that text is converted to a
different encoding.